% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSet.R
\name{eSet.from.fCounts}
\alias{eSet.from.fCounts}
\title{eSet.from.fCounts}
\usage{
eSet.from.fCounts(
  fn.fCounts,
  fCols = cs("Chr Start End Strand Length"),
  maskRemove = "_Aligned.sortedByCoord.out.bam"
)
}
\description{
creates eSet from a \code{featureCounts} output file
}
\details{
\code{featureCounts} is typically run after alignment, it takes bam files as the input and creates a single \code{counts.txt} file
}
