% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-nearest-gene.R
\name{find_nearest_gene}
\alias{find_nearest_gene}
\title{Annotates nearest gene for list of SNPs}
\usage{
find_nearest_gene(data, flanking = 100, build = "hg19",
  collapse = TRUE, snp = "rsid", chr = "chromosome",
  bp = "position")
}
\arguments{
\item{data}{dataframe with results}

\item{flanking}{flanking distance (in kB) on each side (ie. 20kb = spans 40kb)
defaults to 100kb (spans 200kb)}

\item{build}{string specifying build (hg19 (default), hg18, hg38)}

\item{collapse}{if TRUE (default) collapses nearest genes into single commasep.
line. If FALSE returns all annotated genes on separate lines with
a 'distance' column showing how many bp upstream (+) or downstream(-)
the SNP is located}

\item{snp}{snp column name in input data}

\item{chr}{chromosome column name in input data}

\item{bp}{position/bp column name in input data}
}
\value{
data frame with nearest gene specified per marker
}
\description{
Finds the nearest gene for a list of markers. The input data expects a data
frame with headers: rsid, chromosome, position. The list of
annotations are obtained from PLINK documentation
https://www.cog-genomics.org/plink/1.9/resources/
}
