% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_state.R
\name{flexread_counter}
\alias{flexread_counter}
\title{Increment/peek/reset the flexread call counter}
\usage{
flexread_counter(reset = FALSE, peek = FALSE)
}
\description{
Increment/peek/reset the flexread call counter
}
\keyword{internal}
