% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19genelist}
\alias{hg19genelist}
\title{Gene list for hg19}
\format{A data frame with 26292 rows and 4 variables:
\describe{
  \item{CHR}{Chromosome position}
  \item{START}{Start position of gene}
  \item{STOP}{Stop position of gene}
  \item{GENE}{Name of gene}
}}
\usage{
hg19genelist
}
\description{
A dataset containing a list of genes and their bp boundaries for the build
}
\keyword{datasets}
