% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lazyLoadOrCreate}
\alias{lazyLoadOrCreate}
\title{Lazy Load or Build Objects with Optional Caching}
\usage{
lazyLoadOrCreate(...)
}
\description{
Lazily checks for existing objects in the environment or .RData file,
and builds and caches them if not found.
}
\details{
...
}
