% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R, R/rstudio_addinsX.r
\name{selectionTab}
\alias{selectionTab}
\title{Calls tab() for selected object}
\usage{
selectionTab()

selectionTab()
}
\description{
Calls tab() for selected object

Calls tab() for selected object
}
