% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-nearest-gene.R
\name{find_nearest_gene}
\alias{find_nearest_gene}
\title{Annotates nearest gene given input results and genelist. The list of
annotations are stolen from PLINK documentation https://www.cog-genomics.org/plink/1.9/resources}
\usage{
find_nearest_gene(data, flanking, build = "hg19", collapse = TRUE)
}
\arguments{
\item{data}{dataframe with results}

\item{flanking}{flanking distance (in kB) on each side (ie. 20kb = spans 40kb)}

\item{build}{string specifying build (hg19 (default), hg18, hg38)}

\item{collapse}{if TRUE (default) collapses nearest genes into single commasep.
line. If FALSE returns all annotated genes on separate lines with
a 'distance' column showing how many bp upstream (+) or downstream(-)
the SNP is located}
}
\value{
data frame with nearest gene specified per marker
}
\description{
Finds the nearest gene for a list of markers. The input data expects a data
frame with.
}
