% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R
\name{selectionView}
\alias{selectionView}
\title{Explore selected object:
for ExpressionSet, calls exprs(), fData(), and pData()
for table,         calls RStudio's View() function
for function,      calls debugonce()}
\usage{
selectionView()
}
\value{
dashes inside RStudio
}
\description{
Explore selected object:
for ExpressionSet, calls exprs(), fData(), and pData()
for table,         calls RStudio's View() function
for function,      calls debugonce()
}
